BAU: Switch to TreeMapCardInformationStore
Remove the RangeSet-based CardInformationStore

This actually changes behaviour slightly, as we have overlapping BIN ranges!

e.g.

424242424-424242424 is in test-cards

but 424242000-424242999 is in the worldpay live BIN ranges. Who wins?!

In this case, TreeMapCardInformationStore supports both ranges and the
narrowest one will match the card number. So long as one range completely
encloses another, this will work just fine.

So, I think this behavior is technically more correct, even though it is
different to the previous RangeSet behaviour.

As the test cards are not worldpay cards, they do not get processed by the
WorldpayPrepaidParser and so their prepaid status is UNKNOWN.