RED-46: Remove bash usage from the container
We try to run our scripts with bash, but bash isn't installed in the container.

We don't actually need bash, so stop asking for it.

Move the default for JAVA_OPTS to the Dockerfile using `ENV` rather than shell

Run java directly from the Dockerfile without a shell script wrapper. This will
have the side benefit of allowing signals to be delivered directly to the
running Java process, so e.g. Ctrl-c will now kill the application.