Make java PID 1 in the container
Without using the array syntax for `CMD` in the `Dockerfile` we end up
executing `java` with a shell process.

This is unnecessary and undesirable. The shell will absorb the SIGTERM sent to
shut the container down, hiding it from java and leading to an ungraceful
shutdown by the container orchestrator.

Use 'exec' to avoid forking a new process on startup. This will ensure `java`
is process ID 1 within the container and receives signals directly.