REP-190: Stop throwing 500s for short card numbers
If you pass a card number with fewer than 11 characters to cardid, it throws an
internal server error.

e.g.

    curl -v -X POST -H 'Content-Type: application/json' -d '{ "cardNumber": "4"}' http://localhost:8080/v1/api/card

This is because we call substring(0, 11) on the cardnumber string, which throws
a java.lang.StringIndexOutOfBoundsException on strings shorter than 11
characters.

Make this more robust and return 404 instead.

We already test CardInformationStore quite thoroughly so the only functionality
in CardService we need to test is that the lookup key gets truncated
appropriately.

Move this logic to a separate method and test that specifically.