PP-6128 Add data locations to manifest
When running on PaaS Cardid will read the bin range data from a service
named cardid-data. This change adds the env vars which will be set per
space to point to the cardid-data app.

Also, remove defining the route within the manifest since it is set by
terraform when the space is created.