PP-6561 Fix mvn verify phase
The pom references the `dropwizard-dependencies` pom v2.0.10 as a parent which
includes an execution rule for `maven-dependency-plugin:analyze-only`
setting the `failOnWarning` to `true` (the default is false). This plugin is
run as part of the `verify` phase. We have undeclared transitive dependencies so the
`analyze-only` goal fails and consquently the build.

This isn't a problem on Jenkins because it runs `mvn package` for cardid
and not `mvn verify` as per all other apps. By fixing this we can test
all java apps in the same manner.

The result of the pom inheritence can be viewed by running `mvn
help:effective-pom`. Earlier versions of the dropwizard pom
used by connector do not introduce the execution rule with
`failOnWarning` true, though this change may be necessary for those
projects. We should also consider whether we ought to declare transitive
dependencies in our pom.