PP-6561 DO NOT MERGE
Concourse clones cardid using https oauth yet the submodule path is
specified for ssh which we suspect causes problems when the git resource
attempts to clone it as it doesn't have an appropriate private key.

It's possible to use a relative path so that the
submodule should be cloned using the same protocol as that which was
used to clone cardid itself. Trying it out.