PP-8036 Draft changes to use Worldpay DART file
- Delimeter `!` -> `,`
- Row identifier `05` -> `01`
- Card class column 9 -> 10
- Take 11 character substrings of min and max for bin range, as 18 digit
numbers are now given.
- Update card label mappings. Although we could decide to use the
"Scheme product" code in column 13 to provide this mapping.

Remove "prepaidStatus" field from CardInformation as new file specifies
card as either debit, credit or prepaid - so prepaid needs to be a card
type. This will mean we need to change the contract with frontend and
connector change how we determine whether to apply a corporate surcharge
for prepaid corporate cards.