PP-9406 Fix contract test runner
Alter the `url` parameter for the `@PactBroker` annotation to prefix
with "https://", as the the PACT_BROKER_HOST environment variable will
only provide the host name.

Check whether a PactVerificationContext exists in the setup method and
in the TestTemplate method. This will be null if no pacts are found to
verify. This check could be removed once pacts have been uploaded by
connector.