PP-12107: Move BinRangeTest from pay-cardid-data
To make the test work overlapping ranges in the current test data (in
src/main/resources/data-sources) needed to be amended/deleted.

Also removed the long comment in CardIdResourceITest as it claims the
usefulness of it's tests isn't clear. On the contrary the test is useful as it
verifies the API actually works.

Removed `-DskipTests` from the Dockerfile as we want to verify the bin ranges
don't overlap (via running BinRangesTest) whilst building the Docker image.