PP-12107: Move BinRangeTest from pay-cardid-data (#1206)
* PP-12107: Move BinRangeTest from pay-cardid-data

Moving BinRangeTest from pay-cardid-data into this repo as the build-and-push-cardid-candidate CI run will copy the real bin ranges files from s3 into src/main/resources/data-sources. Thus the test will run against real card data in a CI build.

To make the test work overlapping ranges in the current test data (in
src/main/resources/data-sources) needed to be amended/deleted.

Also removed the long comment in CardIdResourceITest as it claims the
usefulness of it's tests isn't clear. On the contrary the test is useful as it
verifies the API actually works.

Removed `-DskipTests` from the Dockerfile as we want to verify the bin ranges
don't overlap (via running BinRangesTest) whilst building the Docker image.