PP-12107: Fix failing test (#1212)
shouldReturnInformationFromBuiltInDefaultBinRangeFiles was failing because in a
build on the CI pipeline, real CSVs from Worldpay are copied to the
src/main/resources/data-sources folder. Thus the card number 1234567890000000
will fail to be found. To get around this, use a well known test card number
4000056655665556 that is also publicly available in
https://docs.stripe.com/testing?locale=en-GB and in the CSV from Worldpay.

This commit also removes "-DskipTests" from the ARM Dockerfile.