PP-14700 Update test BIN range files for American Express test card numbers
- Remove the 37144963500,37144963500 range from the test BIN ranges file. We
believe this range mistakenly exists to support the test card number
371449635398431, which is actually outside of the range but is listed in our
tech docs as a test American Express card number for sandbox and Stripe. The
live BIN ranges file provided to us by Worldpay now contains a range that
overlaps with this range, so we need to remove it from the test BIN ranges
file to prevent a clash. The test card number 371449635398431 will continue
to be supported as a test card number because Stripe still list it as such.
Having a test card number that is defined in the live BIN ranges file rather
than the test BIN ranges file is not ideal but this is the situation we find
ourselves in.

- Move the 34343434343,34343434343 range to the test BIN ranges file. This is
so that we can have the oh-so-memorable test American Express card number of
343434343434343 listed in our tech docs and recognised by sandbox. Worldpay
already list this as a test American Express card number in their docs.